Update custom_domains.py (#100)
* Update custom_domains.py

Added These domains:

arvanstorage.com:
When using `arvancloud.ir`, this domain is constantly requested but it's proxied. this domain should be directed.

digijojeh.com:
The host is in Iran and won't open with proxy.

masirwp.com:
Used alongside with Iranian domain for website-building and hosting.
 
microless.shop:
When opening `microless.ir`, which is an Iranian shopping website hosted in Iran, this domain is constantly requested but because it's proxied the website load time is very high and laggy. This domain should be directed.

boomidi.net:
Hosted in Iran and won't open with a proxy.

shab.cloud:
It's like a CDN for `shab.ir`. with some proxies, images on `shab.ir` won't load. This domain should be directed.

* Update custom_domains.py

Fixed a bug.
The last domain didn't have `,`.

* Update custom_domains.py (Added `kermanmotor.com`)

Added `kermanmotor.com`:
The website won't open with a proxy.